third_party/WebKit/LayoutTests/http/tests/devtools/editor/text-editor-line-breaks-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
This test checks that line endings are inferred from the initial text content, not incremental editing.


Running: testCRInitial
1%0A2%0A3%0A

Running: testCRLFInitial
1%0D%0A2%0D%0A3%0D%0A

Running: testCREdit
1%0Afoo%0Abar2%0A3%0A

Running: testCRLFEdit
1%0D%0Afoo%0D%0Abar2%0D%0A3%0D%0A
